refactor: 최소 단위 컴포넌트 폴더 분리
* 최소 단위 컴포넌트 폴더 분리
