+++ src/pages/SignUp.tsx
@@ -1,11 +1,11 @@
-import { Button } from '../_zlib/components/Button';
-import { Form } from '../_zlib/components/Form';
-import { Icon } from '../_zlib/components/Icon';
-import { Input } from '../_zlib/components/Input';
+import { Button } from '../_zlib/components/common/Button';
+import { Form } from '../_zlib/components/common/Form';
+import { Icon } from '../_zlib/components/common/Icon';
+import { Input } from '../_zlib/components/common/Input';
 import { LineBreak } from '../_zlib/components/LineBreak';
-import { Spacing } from '../_zlib/components/Spacing';
-import { Txt } from '../_zlib/components/Txt';
-import { Wrapper } from '../_zlib/components/Wrapper';
+import { Spacing } from '../_zlib/components/common/Spacing';
+import { Txt } from '../_zlib/components/common/Txt';
+import { Wrapper } from '../_zlib/components/common/Wrapper';
 import useSignUp from '../_zlib/hooks/useSignUp';
 
 export default function SignUp() {
